Move kubetest utils to separate package
