+++ kubetest/gke.go
@@ -33,6 +33,8 @@ import (
 	"strconv"
 	"strings"
 	"time"
+
+	"k8s.io/test-infra/kubetest/util"
 )
 
 const (
@@ -149,7 +151,7 @@ func newGKE(provider, project, zone, region, network, image, cluster string, tes
 	if *gkeCommandGroup == "alpha" || *gkeCommandGroup == "beta" {
 		// By default gcloud {alpha,beta} container is using v1 api.
 		// If we want to use v1alpha1/v1beta1 we need to force it.
-		if err := finishRunning(exec.Command("gcloud", "config", "set", "container/use_v1_api", "False")); err != nil {
+		if err := control.FinishRunning(exec.Command("gcloud", "config", "set", "container/use_v1_api", "False")); err != nil {
 			return nil, err
 		}
 	}
@@ -162,10 +164,10 @@ func newGKE(provider, project, zone, region, network, image, cluster string, tes
 	}
 	g.createCommand = append(g.createCommand, createArgs...)
 
-	if err := migrateOptions([]migratedOption{{
-		env:    "CLOUDSDK_API_ENDPOINT_OVERRIDES_CONTAINER",
-		option: gkeEnvironment,
-		name:   "--gke-environment",
+	if err := util.MigrateOptions([]util.MigratedOption{{
+		Env:    "CLOUDSDK_API_ENDPOINT_OVERRIDES_CONTAINER",
+		Option: gkeEnvironment,
+		Name:   "--gke-environment",
 	}}); err != nil {
 		return nil, err
 	}
@@ -230,7 +232,7 @@ func newGKE(provider, project, zone, region, network, image, cluster string, tes
 	// set --num-nodes flag for ginkgo, since NUM_NODES is not set for gke deployer.
 	numNodes := strconv.Itoa(g.shape[defaultPool].Nodes)
 	// testArgs can be empty, and we need to support this case
-	*testArgs = strings.Join(setFieldDefault(strings.Fields(*testArgs), "--num-nodes", numNodes), " ")
+	*testArgs = strings.Join(util.SetFieldDefault(strings.Fields(*testArgs), "--num-nodes", numNodes), " ")
 
 	if *upgradeArgs != "" {
 		// --upgrade-target will be passed to e2e upgrade framework to get a valid update version.
@@ -238,7 +240,7 @@ func newGKE(provider, project, zone, region, network, image, cluster string, tes
 		// Here we special case for gke-latest and will extract an actual valid gke version.
 		// - gke-latest will be resolved to the latest gke version, and
 		// - gke-latest-1.7 will be resolved to the latest 1.7 patch version supported on gke.
-		fields, val, exist := extractField(strings.Fields(*upgradeArgs), "--upgrade-target")
+		fields, val, exist := util.ExtractField(strings.Fields(*upgradeArgs), "--upgrade-target")
 		if exist {
 			if strings.HasPrefix(val, "gke-latest") {
 				releasePrefix := ""
@@ -249,9 +251,9 @@ func newGKE(provider, project, zone, region, network, image, cluster string, tes
 					return nil, fmt.Errorf("fail to get latest gke version : %v", err)
 				}
 			}
-			fields = setFieldDefault(fields, "--upgrade-target", val)
+			fields = util.SetFieldDefault(fields, "--upgrade-target", val)
 		}
-		*upgradeArgs = strings.Join(setFieldDefault(fields, "--num-nodes", numNodes), " ")
+		*upgradeArgs = strings.Join(util.SetFieldDefault(fields, "--num-nodes", numNodes), " ")
 	}
 
 	g.singleZoneNodeInstanceGroup = *gkeSingleZoneNodeInstanceGroup
@@ -261,11 +263,11 @@ func newGKE(provider, project, zone, region, network, image, cluster string, tes
 
 func (g *gkeDeployer) Up() error {
 	// Create network if it doesn't exist.
-	if err := finishRunning(exec.Command("gcloud", "compute", "networks", "describe", g.network,
+	if err := control.FinishRunning(exec.Command("gcloud", "compute", "networks", "describe", g.network,
 		"--project="+g.project,
 		"--format=value(name)")); err != nil {
 		// Assume error implies non-existent.
-		if err := finishRunning(exec.Command("gcloud", "compute", "networks", "create", g.network,
+		if err := control.FinishRunning(exec.Command("gcloud", "compute", "networks", "create", g.network,
 			"--project="+g.project,
 			"--mode=auto")); err != nil {
 			return err
@@ -277,7 +279,7 @@ func (g *gkeDeployer) Up() error {
 		createSubnetCommand := []string{"compute", "networks", "subnets", "create"}
 		createSubnetCommand = append(createSubnetCommand, "--project="+g.project, "--network="+g.network)
 		createSubnetCommand = append(createSubnetCommand, customSubnetFields...)
-		if err := finishRunning(exec.Command("gcloud", createSubnetCommand...)); err != nil {
+		if err := control.FinishRunning(exec.Command("gcloud", createSubnetCommand...)); err != nil {
 			return err
 		}
 		g.subnetwork = customSubnetFields[0]
@@ -311,14 +313,14 @@ func (g *gkeDeployer) Up() error {
 		args = append(args, "--cluster-version="+v)
 	}
 	args = append(args, g.cluster)
-	if err := finishRunning(exec.Command("gcloud", args...)); err != nil {
+	if err := control.FinishRunning(exec.Command("gcloud", args...)); err != nil {
 		return fmt.Errorf("error creating cluster: %v", err)
 	}
 	for poolName, pool := range g.shape {
 		if poolName == defaultPool {
 			continue
 		}
-		if err := finishRunning(exec.Command("gcloud", g.containerArgs(
+		if err := control.FinishRunning(exec.Command("gcloud", g.containerArgs(
 			"node-pools", "create", poolName,
 			"--cluster="+g.cluster,
 			"--project="+g.project,
@@ -385,7 +387,7 @@ export KUBERNETES_PROVIDER=gke
 	} else {
 		dumpCmd = fmt.Sprintf("./cluster/log-dump/log-dump.sh '%s' '%s'", localPath, gcsPath)
 	}
-	return finishRunning(exec.Command("bash", "-c", fmt.Sprintf(gkeLogDumpTemplate,
+	return control.FinishRunning(exec.Command("bash", "-c", fmt.Sprintf(gkeLogDumpTemplate,
 		g.project,
 		g.zone,
 		strings.Join(filters, " OR "),
@@ -425,7 +427,7 @@ func (g *gkeDeployer) getKubeConfig() error {
 	if err := os.Setenv("KUBECONFIG", g.kubecfg); err != nil {
 		return err
 	}
-	if err := finishRunning(exec.Command("gcloud", g.containerArgs("clusters", "get-credentials", g.cluster,
+	if err := control.FinishRunning(exec.Command("gcloud", g.containerArgs("clusters", "get-credentials", g.cluster,
 		"--project="+g.project,
 		g.location)...)); err != nil {
 		return fmt.Errorf("error executing get-credentials: %v", err)
@@ -459,7 +461,7 @@ func (g *gkeDeployer) ensureFirewall() error {
 	if err != nil {
 		return fmt.Errorf("error getting unique firewall: %v", err)
 	}
-	if finishRunning(exec.Command("gcloud", "compute", "firewall-rules", "describe", firewall,
+	if control.FinishRunning(exec.Command("gcloud", "compute", "firewall-rules", "describe", firewall,
 		"--project="+g.project,
 		"--format=value(name)")) == nil {
 		// Assume that if this unique firewall exists, it's good to go.
@@ -472,14 +474,14 @@ func (g *gkeDeployer) ensureFirewall() error {
 		"--limit=1",
 		"--format=get(tags.items)").Output()
 	if err != nil {
-		return fmt.Errorf("instances list failed: %s", execError(err))
+		return fmt.Errorf("instances list failed: %s", util.ExecError(err))
 	}
 	tag := strings.TrimSpace(string(tagOut))
 	if tag == "" {
 		return fmt.Errorf("instances list returned no instances (or instance has no tags)")
 	}
 
-	if err := finishRunning(exec.Command("gcloud", "compute", "firewall-rules", "create", firewall,
+	if err := control.FinishRunning(exec.Command("gcloud", "compute", "firewall-rules", "create", firewall,
 		"--project="+g.project,
 		"--network="+g.network,
 		"--allow="+e2eAllow,
@@ -498,7 +500,7 @@ func (g *gkeDeployer) getInstanceGroups() error {
 		"--project="+g.project,
 		g.location)...).Output()
 	if err != nil {
-		return fmt.Errorf("instance group URL fetch failed: %s", execError(err))
+		return fmt.Errorf("instance group URL fetch failed: %s", util.ExecError(err))
 	}
 	igURLs := strings.Split(strings.TrimSpace(string(igs)), ";")
 	if len(igURLs) == 0 {
@@ -535,18 +537,18 @@ func (g *gkeDeployer) Down() error {
 	g.instanceGroups = nil
 
 	// We best-effort try all of these and report errors as appropriate.
-	errCluster := finishRunning(exec.Command(
+	errCluster := control.FinishRunning(exec.Command(
 		"gcloud", g.containerArgs("clusters", "delete", "-q", g.cluster,
 			"--project="+g.project,
 			g.location)...))
-	errFirewall := finishRunning(exec.Command("gcloud", "compute", "firewall-rules", "delete", "-q", firewall,
+	errFirewall := control.FinishRunning(exec.Command("gcloud", "compute", "firewall-rules", "delete", "-q", firewall,
 		"--project="+g.project))
 	var errSubnet error
 	if g.subnetwork != "" {
-		errSubnet = finishRunning(exec.Command("gcloud", "compute", "networks", "subnets", "delete", "-q", g.subnetwork,
+		errSubnet = control.FinishRunning(exec.Command("gcloud", "compute", "networks", "subnets", "delete", "-q", g.subnetwork,
 			g.subnetworkRegion, "--project="+g.project))
 	}
-	errNetwork := finishRunning(exec.Command("gcloud", "compute", "networks", "delete", "-q", g.network,
+	errNetwork := control.FinishRunning(exec.Command("gcloud", "compute", "networks", "delete", "-q", g.network,
 		"--project="+g.project))
 	if errCluster != nil {
 		return fmt.Errorf("error deleting cluster: %v", errCluster)
@@ -568,7 +570,7 @@ func (g *gkeDeployer) containerArgs(args ...string) []string {
 }
 
 func (g *gkeDeployer) GetClusterCreated(gcpProject string) (time.Time, error) {
-	res, err := output(exec.Command(
+	res, err := control.Output(exec.Command(
 		"gcloud",
 		"compute",
 		"instance-groups",
