+++ kubetest/process/BUILD.bazel
@@ -1,0 +1,30 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["process.go"],
+    importpath = "k8s.io/test-infra/kubetest/process",
+    visibility = ["//visibility:public"],
+    deps = ["//kubetest/util:go_default_library"],
+)
+
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [":package-srcs"],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["process_test.go"],
+    embed = [":go_default_library"],
+    deps = ["//kubetest/util:go_default_library"],
+)
